controle

+++ src/aplicacao/controle/AgendamentoActionListener.java
@@ -7,6 +7,7 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
 import java.util.List;
+import javax.swing.JOptionPane;
 
 public class AgendamentoActionListener implements ActionListener{
     private CadastroAgendamento frame;
@@ -26,6 +27,18 @@ public class AgendamentoActionListener implements ActionListener{
             case "Fechar":
                 frame.dispose();
             break;
+            case "Limpar":
+                limpaCampos();
+            break;
+            case "Buscar":
+                if(dao.existis(a.getCodAgend())){
+                    Agendamento nv  = new Agendamento();
+                    nv = dao.getAgendamento(a.getCodAgend());
+                    frame.setAgendamento(nv);
+                } else {
+                    JOptionPane.showMessageDialog(frame, "Não Há Resultados Para Sua Busca", "Atenção!!!", 0);
+                }
+            break;
             case "Salvar":
                 if(!dao.existis(a.getCodAgend())){
                     dao.insert(a);
@@ -34,8 +47,30 @@ public class AgendamentoActionListener implements ActionListener{
                 }
             break;
             case "Excluir":
-                dao.delete(a);
+                if(dao.existis(a.getCodAgend())){
+                    dao.delete(a);
+                    limpaCampos();
+                } else {
+                    JOptionPane.showMessageDialog(frame, "Código Inexistente", "Atenção!!!", 0);
+                }
             break;
         }
+    }
+    
+    private void limpaCampos(){
+        Agendamento a  = new Agendamento();
+        
+        a.setCodAgend(0);
+        a.setStatus(0);
+        a.setDataIni("");
+        a.setHoraIni("");
+        a.setKmIni(0);
+        a.setDataFin("");
+        a.setHorafin("");
+        a.setKmFin(0);
+        a.setCodMot(0);
+        a.setCodVei(0);
+        a.setObs("");
+        frame.setAgendamento(a);
     }
 }
