+++ src/aplicacao/controle/MotoristaActionListener.java
@@ -3,10 +3,12 @@ package aplicacao.controle;
 
 import aplicacao.view.CadastroMotorista;
 import aplicacao.modelo.Motorista;
+import aplicacao.modelo.Veiculo;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
 import java.util.List;
+import javax.swing.JOptionPane;
 
 public class MotoristaActionListener implements ActionListener{
     private CadastroMotorista frame;
@@ -26,6 +28,18 @@ public class MotoristaActionListener implements ActionListener{
             case "Fechar":
                 frame.dispose();
             break;
+            case "Limpar":
+                limpaCampos();
+            break;
+            case "Buscar":
+                if(dao.existis(m.getCodMot())){
+                    Motorista nv  = new Motorista();
+                    nv = dao.getMotorista(m.getCodMot());
+                    frame.setMotorista(nv);
+                } else {
+                    JOptionPane.showMessageDialog(frame, "Não Há Resultados Para Sua Busca", "Atenção!!!", 0);
+                }
+            break;
             case "Salvar":
                 if(!dao.existis(m.getCodMot())){
                     dao.insert(m);
@@ -34,8 +48,24 @@ public class MotoristaActionListener implements ActionListener{
                 }
             break;
             case "Excluir":
-                dao.delete(m);
+                if(dao.existis(m.getCodMot())){
+                    dao.delete(m);
+                    limpaCampos();
+                } else {
+                    JOptionPane.showMessageDialog(frame, "Código Inexistente", "Atenção!!!", 0);
+                }
             break;
         }
+    }
+    
+    private void limpaCampos(){
+        Motorista m  = new Motorista();
+        
+        m.setCodMot(0);
+        m.setNome("");
+        m.setTipoCNH(0);
+        m.setCNH(0);
+        m.setValCNH("");
+        frame.setMotorista(m);
     }
 }
